# public/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: PFE E-commerce API
  description: API documentation.
  version: 1.0.0
servers:
  - url: /api
    description: API Server
paths:
...
  /profile/verify/send:
    post:
      summary: Send verification email
      description: Sends a verification link to the user email. Requires a valid Bearer token. Returns 400 if email is already verified.
      operationId: sendVerificationEmail
      tags:
        - Email Verification
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification link sent successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "success"
                message: "Lien envoye ! Verifiez votre boite mail"
        '400':
          description: Email already verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "error"
                message: "Email deja verifie"
        '401':
          description: Unauthenticated (token missing or invalid).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /verify/{id}/{hash}:
    get:
      summary: Verify email
      description: Validates the user email using the signed link received by email.
      operationId: verifyEmail
      tags:
        - Email Verification
      parameters:
        - name: id
          in: path
          required: true
          description: The user ID.
          schema:
            type: integer
            example: 1
        - name: hash
          in: path
          required: true
          description: The SHA1 hash of the user email.
          schema:
            type: string
            example: "a94a8fe5ccb19ba61c4c0873d391e987982fbbd3"
      responses:
        '200':
          description: Email verified successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "success"
                message: "E-mail valide avec succes !"
        '400':
          description: Invalid link or email already verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "error"
                message: "Lien invalide"
  
   # ========== PASSWORD RESET ==========
  /password/email:
    post:
      summary: Demander un lien de réinitialisation
      description: Envoie un lien de réinitialisation du mot de passe à l'adresse email fournie. Retourne toujours 200 même si l'email n'existe pas (sécurité anti-énumération).
      operationId: sendResetLink
      tags:
        - Password Reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordEmailRequest'
            example:
              email: "jamal@example.com"
      responses:
        '200':
          description: Lien envoyé (ou email inexistant — même réponse pour la sécurité).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "success"
                message: "Si cet email existe, un lien de réinitialisation a été envoyé."
        '422':
          description: Validation error (email manquant ou format invalide).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /password/reset/{token}:
    get:
      summary: Valider le token de réinitialisation
      description: Vérifie que le token reçu par email est valide et non expiré (60 minutes). Retourne l'email et le token si valide.
      operationId: validateToken
      tags:
        - Password Reset
      parameters:
        - name: token
          in: path
          required: true
          description: Le token de 64 caractères reçu dans le lien email.
          schema:
            type: string
            example: "aB3dEfGhIjKlMnOpQrStUvWxYz1234567890aB3dEfGhIjKlMnOpQrStUvWx"
        - name: email
          in: query
          required: true
          description: L'adresse email de l'utilisateur.
          schema:
            type: string
            format: email
            example: "mohamed@example.com"
      responses:
        '200':
          description: Token valide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
        '400':
          description: Token invalide, expiré, ou email manquant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "error"
                message: "Lien invalide ou expiré."
  /password/update:
    post:
      summary: Mettre à jour le mot de passe
      description: Met à jour le mot de passe de l'utilisateur après validation du token. Révoque tous les tokens Sanctum existants. Si l'email n'était pas vérifié, il est marqué comme vérifié.
      operationId: updatePassword
      tags:
        - Password Reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdateRequest'
            example:
              email: "jamal@example.com"
              token: "FIc64WxXBDdnkGPgnF8PzCWqZGvFCPJ1WHzfoLtMWrQ5Cb2qolDfZajGmKTMV5en"
              password: "NouveauMdp1!"
              c_password: "NouveauMdp1!"
      
      responses:
        '200':
          description: Mot de passe mis à jour avec succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "success"
                message: "Mot de passe mis à jour avec succès. Veuillez vous reconnecter."
        '400':
          description: Token invalide ou expiré.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "error"
                message: "Lien invalide ou expiré."
        '404':
          description: Utilisateur introuvable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                status: "error"
                message: "Utilisateur introuvable."
        '422':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  # ========== CATALOGUE PRODUITS ==========
...
  /orders/{id}/download-receipt:
    get:
      summary: Télécharger le reçu PDF
      description: Télécharge le reçu PDF d'une commande spécifique.
      operationId: downloadReceipt
      tags:
        - Commandes
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: L'ID de la commande.
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Fichier PDF téléchargé.
        '404':
          description: Commande ou reçu introuvable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
components:
  schemas:
...
    PasswordEmailRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "jamal@example.com"
    PasswordUpdateRequest:
      type: object
      required:
        - email
        - token
        - password
        - c_password
      properties:
        email:
          type: string
          format: email
          example: "jamal@example.com"
        token:
          type: string
          example: "FIc64WxXBDdnkGPgnF8PzCWqZGvFCPJ1WHzfoLtMWrQ5Cb2qolDfZajGmKTMV5en"
        password:
          type: string
          format: password
          example: "NouveauMdp1!"
        c_password:
          type: string
          example: "NouveauMdp1!"
...
    ValidateTokenResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Token valide. Vous pouvez maintenant choisir un nouveau mot de passe."
        email:
          type: string
          example: "mohamed@example.com"
        token:
          type: string
          example: "aB3dEfGhIjKlMnOpQrStUvWxYz1234567890aB3dEfGhIjKlMnOpQrStUvWx"
...
    MessageResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Email ou mot de passe incorrect"
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            nom: ["The combination of this Nom and Prenom already exists."]
            email: ["The email has already been taken."]
            password: ["The password field must contain at least one symbol."]

#produits et categories    
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT